added scripts for creating copies of test database

+++ scripts/test-setup.sh
@@ -2,16 +2,15 @@
 
 set -e
 
-source $WORKSPACE/scripts/parse-env.sh
+if ! [ -n "$GIT_BRANCH" ]
+then
+	echo "Error: GIT_BRANCH is not defined!"
+	exit 1
+fi
+
+source $WORKSPACE/scripts/parse-env.sh $GIT_BRANCH
 
 $git reset --hard
 
-echo Copying database $DBSOURCE to $DBNAME
-$mysql -e 'DROP DATABASE IF EXISTS '$DBNAME';'
-$mysql -e 'CREATE DATABASE '$DBNAME';'
-time $mysqldump --opt $DBSOURCE | $mysql $DBNAME
-
-sed -i -e 's/'$DBSOURCE'/'$DBNAME'/g' \
-	-e 's/unifina-test/root/g' \
-	-e 's/password.*/password="'$MYSQL_PW'"/' $DSCONFIG
-
+time $WORKSPACE/scripts/copy-test-db.sh $GIT_BRANCH
+sed -i -e 's/'$DBSOURCE'/'$DBNAME'/g' $DSCONFIG
