+++ scripts/test-teardown.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
-source $WORKSPACE/scripts/parse-env.sh
+if ! [ -n "$GIT_BRANCH" ]
+then
+	echo "Error: GIT_BRANCH is not defined!"
+	exit 1
+fi
+
+source $WORKSPACE/scripts/parse-env.sh $GIT_BRANCH
 
 # drop test db
 $mysql -e 'DROP DATABASE IF EXISTS '$DBNAME';'
 
 $git checkout $DSCONFIG
-$git checkout $BUILDCONFIG
-
